Made sections more modular
created a file for header, footer, and head sections and put them (and
Menu) into a section directory. Modified all files so that their
header,footer,head, and menu sections referenced the above .php files so
that all changes are made to one file instead of making the same changes
to multiple files.